TST: add test for getting array from series
